requirements.txt
# =============================================================================
# FinTech Fraud Detection Pipeline
# =============================================================================

# Kafka Producer
kafka-python==2.0.2
confluent-kafka==2.3.0

# Apache Spark (PySpark)
pyspark==3.5.0

# PostgreSQL
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Apache Airflow
apache-airflow==2.7.3

# Data Processing
pandas==2.1.3
pyarrow==14.0.1

# Utilities
python-dateutil==2.8.2
faker==20.1.0
